Separate out application start from routing table
